+++ sdk/metric/exemplar.go
@@ -6,11 +6,15 @@ package metric // import "go.opentelemetry.io/otel/sdk/metric"
 import (
 	"os"
 	"runtime"
-	"slices"
 
 	"go.opentelemetry.io/otel/sdk/metric/exemplar"
 	"go.opentelemetry.io/otel/sdk/metric/internal/aggregate"
 )
+
+// ExemplarReservoirProviderSelector selects the
+// [go.opentelemetry.io/otel/sdk/metric/exemplar.ReservoirProvider] to use
+// based on the [Aggregation] of the metric.
+type ExemplarReservoirProviderSelector func(Aggregation) exemplar.ReservoirProvider
 
 // reservoirFunc returns the appropriately configured exemplar reservoir
 // creation func based on the passed InstrumentKind and user defined
@@ -36,16 +40,35 @@ func reservoirFunc[N int64 | float64](agg Aggregation) func() aggregate.Filtered
 		filter = exemplar.SampledFilter
 	}
 
+	provider := DefaultExemplarReservoirProviderSelector(agg)
+	return func() aggregate.FilteredExemplarReservoir[N] {
+		return aggregate.NewFilteredExemplarReservoir[N](filter, provider())
+	}
+}
+
+// DefaultExemplarReservoirProviderSelector returns the default
+// [go.opentelemetry.io/otel/sdk/metric/exemplar.ReservoirProvider] for the
+// provided [Aggregation].
+//
+// For explicit bucket histograms with more than 1 bucket, it uses the
+// [go.opentelemetry.io/otel/sdk/metric/exemplar.HistogramReservoirProvider].
+// For exponential histograms, it uses the
+// [go.opentelemetry.io/otel/sdk/metric/exemplar.FixedSizeReservoirProvider]
+// with a size of min(20, max_buckets).
+// For all other aggregations, it uses the
+// [go.opentelemetry.io/otel/sdk/metric/exemplar.FixedSizeReservoirProvider]
+// with a size equal to the number of CPUs.
+//
+// Exemplar default reservoirs MAY change in a minor version bump. No
+// guarantees are made on the shape or statistical properties of returned
+// exemplars.
+func DefaultExemplarReservoirProviderSelector(agg Aggregation) exemplar.ReservoirProvider {
 	// https://github.com/open-telemetry/opentelemetry-specification/blob/d4b241f451674e8f611bb589477680341006ad2b/specification/metrics/sdk.md#exemplar-defaults
 	// Explicit bucket histogram aggregation with more than 1 bucket will
 	// use AlignedHistogramBucketExemplarReservoir.
 	a, ok := agg.(AggregationExplicitBucketHistogram)
 	if ok && len(a.Boundaries) > 0 {
-		cp := slices.Clone(a.Boundaries)
-		return func() aggregate.FilteredExemplarReservoir[N] {
-			bounds := cp
-			return aggregate.NewFilteredExemplarReservoir[N](filter, exemplar.NewHistogramReservoir(bounds))
-		}
+		return exemplar.HistogramReservoirProvider(a.Boundaries)
 	}
 
 	var n int
@@ -72,7 +95,5 @@ func reservoirFunc[N int64 | float64](agg Aggregation) func() aggregate.Filtered
 		}
 	}
 
-	return func() aggregate.FilteredExemplarReservoir[N] {
-		return aggregate.NewFilteredExemplarReservoir[N](filter, exemplar.NewFixedSizeReservoir(n))
-	}
+	return exemplar.FixedSizeReservoirProvider(n)
 }